Implementing galaxy rotation
Former-commit-id: 196a595d2e5f5727923c36169d4d32db6402dea8

+++ Renderer/Renderer/Mesh.cpp
@@ -36,7 +36,7 @@ void Mesh::Draw(const Shader & shader) const
         else if (name == "texture_specular")
             number = std::to_string(specular_num++);
 
-        shader.SetFloat((name + number).c_str(), i);
+        shader.SetFloat((name + number).c_str(), static_cast<float>(i));
         glBindTexture(GL_TEXTURE_2D, textures[i].id);
     }
     glActiveTexture(GL_TEXTURE0);
